Added recorded test data. Made a separate test case for Vulnerability API.
