+++ tests/VulnerabilityAlerts.py
@@ -1,0 +1,20 @@
+import Framework
+import github
+import datetime
+
+class VulnerabilityAlerts(Framework.TestCase):
+    def setUp(self):
+        Framework.TestCase.setUp(self)
+        self.user = self.g.get_user()
+        self.repo = self.user.get_repo("PyGithub")
+    
+
+    def testGetStatusOfVulnerabilityAlerts(self):
+        self.assertTrue(self.repo.get_vulnerability_alert_status())
+
+    def testEnableVulnerabilityAlerts(self):
+        self.assertTrue(self.repo.enable_vulnerability_alerts())
+
+    
+    def testDisableVulnerabilityAlerts(self):
+        self.assertTrue(self.repo.disable_vulnerability_alerts())
